refactored the home page
