+++ frontend/src/pages/index/components/main/rainbowText.tsx
